requirements.txt
python-telegram-bot==11.1.0
PyYAML==5.1.2
RPi.GPIO==0.7.0
